# swagger.yaml
basePath: /
consumes:
- text/plain
definitions:
  addBloodMsgReq:
    properties:
      AddMsgCateID:
        format: uint32
        type: integer
      AngX:
        format: float
        type: number
      AngY:
        format: float
        type: number
      AngZ:
        format: float
        type: number
      BlockID:
        format: uint32
        type: integer
      CharacterID:
        type: string
      MainMsgID:
        format: uint32
        type: integer
      MsgID:
        format: uint32
        type: integer
      PosX:
        format: float
        type: number
      PosY:
        format: float
        type: number
      PosZ:
        format: float
        type: number
      Version:
        format: int64
        type: integer
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
...
  deleteBloodMsgReq:
    properties:
      BloodMsgID:
        format: int64
        type: integer
      Version:
        format: int64
        type: integer
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
  finaliseMultiplayReq:
    properties:
      CharacterID:
        type: string
      GradeA:
        format: int64
        type: integer
      GradeB:
        format: int64
        type: integer
      GradeC:
        format: int64
        type: integer
      GradeD:
        format: int64
        type: integer
      GradeS:
        format: int64
        type: integer
      Version:
        format: int64
        type: integer
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
  getBloodMsgReq:
    properties:
      BlockID:
        format: uint32
        type: integer
      CharacterID:
        type: string
      ReplayNum:
        format: int64
        type: integer
      Version:
        format: int64
        type: integer
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
...
  initMultiplayHandler:
    properties:
      CharacterID:
        type: string
      Version:
        format: int64
        type: integer
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
...
  outOfBlockReq:
    properties:
      CharacterID:
        type: string
      Version:
        format: int64
        type: integer
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
...
  updateBloodMsgGradeReq:
    properties:
      BloodMsgID:
        format: int64
        type: integer
      Version:
        format: int64
        type: integer
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
  updateOtherPlayerGradeReq:
    properties:
      CharacterID:
        type: string
      Grade:
        format: int64
        type: integer
      Version:
        format: int64
        type: integer
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
...
info:
  description: The boostrap and game API endpoints for Demon's Souls
  title: Demon's Souls API
  version: 1.0.0
paths:
...
  /cgi-bin/addBloodMessage.spd:
    post:
      consumes:
      - text/plain
      description: Adds a new blood message
      operationId: addBloodMsgHandler
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/addBloodMsgReq'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: Add blood message
      tags:
      - messages
...
  /cgi-bin/deleteBloodMessage.spd:
    post:
      consumes:
      - text/plain
      description: Deletes a new blood message
      operationId: deleteBloodMsgHandler
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/deleteBloodMsgReq'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: Delete blood message
      tags:
      - messages
  /cgi-bin/finalizeMultiPlay.spd:
    post:
      consumes:
      - text/plain
      operationId: finaliseMultiplayHandler
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/finaliseMultiplayReq'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: Finalise multiplayer
      tags:
      - multiplayer
  /cgi-bin/getBloodMessage.spd:
    post:
      consumes:
      - text/plain
      description: Returns a list of blood messages within a given area of the game
      operationId: getBloodMsgHandler
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/getBloodMsgReq'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: Get blood messages
      tags:
      - messages
...
  /cgi-bin/initializeMultiPlay.spd:
    post:
      consumes:
      - text/plain
      description: Initialise multiplayer for a given character
      operationId: initMultiplayHandler
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/initMultiplayHandler'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: Initialise multiplayer
      tags:
      - multiplayer
...
  /cgi-bin/outOfBlock.spd:
    post:
      consumes:
      - text/plain
      description: Triggered when a player leaves an area of the game
      operationId: outOfBlockHandler
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/outOfBlockReq'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: Out of block
      tags:
      - multiplayer
...
  /cgi-bin/updateBloodMessageGrade.spd:
    post:
      consumes:
      - text/plain
      description: Updates the grade of a blood message
      operationId: updateBloodMsgGradeHandler
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/updateBloodMsgGradeReq'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: Update blood message grade
      tags:
      - messages
  /cgi-bin/updateOtherPlayerGrade.spd:
    post:
      consumes:
      - text/plain
      description: Sets the grade of a character after a multiplayer session
      operationId: updateOtherPlayerGradeHandler
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/updateOtherPlayerGradeReq'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: Update other character grade
      tags:
      - multiplayer
produces:
- text/plain
schemes:
- http
swagger: "2.0"